Function that adds a new node at the end of a list

+++ 0x12-singly_linked_lists/3-add_node_end.c
@@ -6,14 +6,15 @@
 /**
  * add_node - adds new node at the end of a list
  * @str: string to be pointed to
- * head: pointer to the address copied
+ * @head: pointer to the address copied
  * Return: address of new element or Null if fails
  */
 
 list_t *add_node_end(list_t **head, const char *str)
 {
-	unsigned int len = 0;
+	size_t len = 0;
 	list_t *newnode, *traverse;
+
 	newnode = malloc(sizeof(list_t));
 
 	for (; str[len]; len++)
